requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cos-python-sdk-v5==1.8.0
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==3.0.5
django-import-export==2.0.2
django-simpleui==4.0
docutils==0.16
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
idna==2.9
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.2
joblib==0.14.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.1
odfpy==1.4.1
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.0.0
prettytable==0.7.2
pyecharts==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tablib==1.1.0
torch @ file:///E:/Program/Service-Outsourcing/mod/torch-1.4.0-cp37-cp37m-win_amd64.whl
torchvision==0.5.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
